chore(cli): fold adapters into COMMANDS and usage (WM-850)
Move the read-only adapters verb off the inline cli.mjs dispatch override so it lives with the other commands.
